wip, load the unmerged file when merging lane from scope
