+++ components/legacy/extension-data/extension-data-list.ts
@@ -7,6 +7,8 @@ import {
   reStructureBuildArtifacts,
 } from '@teambit/component.sources';
 import { ExtensionDataEntry, REMOVE_EXTENSION_SPECIAL_SIGN } from './extension-data';
+import EnvsAspect from '@teambit/envs';
+import DependencyResolverAspect from '@teambit/dependency-resolver';
 
 type ExtensionConfig = { [extName: string]: any } | RemoveExtensionSpecialSign;
 type ConfigOnlyEntry = {
@@ -148,6 +150,39 @@ export class ExtensionDataList extends Array<ExtensionDataEntry> {
     return ExtensionDataList.fromArray(arr);
   }
 
+  /**
+   * from the merge-config we get the dep-resolver policy as an array, because it needs to support "force" prop.
+   * however, when we save the config, we want to save it as an object, so we need split the data into two:
+   * 1. force: true, which gets saved into the config.
+   * 2. force: false, which gets saved into the data.dependencies later on. see the workspace.getAutoDetectOverrides()
+   */
+  removeAutoDepsFromConfig() {
+    const policy = this.findCoreExtension(DependencyResolverAspect.id)?.config.policy;
+    if (!policy) return;
+
+    const autoDepsObj = {};
+    ['dependencies', 'devDependencies', 'peerDependencies'].forEach((key) => {
+      if (!policy[key]) return;
+      // this is only relevant when it is saved as an array. otherwise, it's always force: true.
+      if (!Array.isArray(policy[key])) return;
+
+      autoDepsObj[key] = policy[key].filter((dep) => !dep.force);
+      policy[key] = policy[key].filter((dep) => dep.force);
+
+      if (!policy[key].length) {
+        delete policy[key];
+        return;
+      }
+      // convert to object
+      policy[key] = policy[key].reduce((acc, current) => {
+        acc[current.name] = current.version;
+        return acc;
+      }, {});
+    });
+
+    return autoDepsObj;
+  }
+
   static fromConfigObject(obj: { [extensionId: string]: any } = {}): ExtensionDataList {
     const arr = Object.keys(obj)
       // We don't want to store extensions with the file protocol because they are bounded to a specific machine.
@@ -172,11 +207,6 @@ export class ExtensionDataList extends Array<ExtensionDataEntry> {
    * Make sure you extension ids are resolved before call this, otherwise you might get unexpected results
    * for example:
    * you might have 2 entries like: default-scope/my-extension and my-extension on the same time
-   *
-   * @static
-   * @param {ExtensionDataList[]} list
-   * @returns {ExtensionDataList}
-   * @memberof ExtensionDataList
    */
   static mergeConfigs(list: ExtensionDataList[], ignoreVersion = true): ExtensionDataList {
     if (list.length === 1) {
@@ -187,6 +217,17 @@ export class ExtensionDataList extends Array<ExtensionDataEntry> {
 
     const merged = list.reduce(mergeReducer, new ExtensionDataList());
     return ExtensionDataList.fromArray(merged);
+  }
+
+  /**
+   * makes sure the "env" prop has the id without a version, and the full env-id is in the root of the object
+   */
+  static adjustEnvsOnConfigObject(conf: Record<string, any>) {
+    const env = conf[EnvsAspect.id]?.env;
+    if (!env) return;
+    const [id] = env.split('@');
+    conf[EnvsAspect.id] = { env: id };
+    conf[env] = {};
   }
 }
